// src/models/file-model.ts
export interface FileModel {
  alternativeText?: string
  caption?: string
  created_at?: string
  ext?: string
  formats?: string
  hash?: string
  height?: string
  id?: any
  mime?: string
  name?: string
  previewUrl?: string
  provider?: string
  provider_metadata?: string
  related?: any[]
  size?: number
  updated_at?: string
  url?: string
  width?: number
}